+++ client/changes.md
@@ -136,9 +136,8 @@ if (createdBy === me || isEditable) allowEdit
 - ~~fix file upload error (JSON PARSING) --~~
 - ~~fix edit task shared by user --~~
 
-
 ### 27-2-2024
 
-- move all business logic to service
-- also check if task is editable if it's being edited by shared with
+- move all business logic to service --
+- also check if task is editable if it's being edited by shared with --
 - loading indicators...